Fixed some rendering issues, and fixed writing/reading values to/from camera.

+++ src/cameraproperty.cpp
@@ -3,6 +3,6 @@
 
 using namespace CameraManager;
 
-CameraProperty::CameraProperty(CameraManager::PropertyType type, QString name, float min, float max, int decimals, bool canAuto, bool isAuto, bool onOff, float value)
+CameraProperty::CameraProperty(CameraManager::PropertyType type, QString name, double min, double max, int decimals, bool canAuto, bool isAuto, bool onOff, double value)
     : type(type), name(name), min(min), max(max), decimals(decimals), canAuto(canAuto), isAuto(isAuto), onOff(onOff), val(value) {
 }
