+++ src/cameraproperty.h
@@ -43,7 +43,7 @@ namespace CameraManager {
          * @param decimals number of decimal for that property (used for the slider)
          * @param canAuto true if that property can be set to auto (will desactivate the slider)
          */  
-        CameraProperty(PropertyType type, QString name, float min, float max, int decimals = 0, bool canAuto = false, bool isAuto = true, bool onOff = true, float value = 0);  // Lars Aksel 29.01.2015
+        CameraProperty(PropertyType type, QString name, double min, double max, int decimals = 0, bool canAuto = false, bool isAuto = true, bool onOff = true, double value = 0);  // Lars Aksel 29.01.2015
 
 	    PropertyType getType(){ return type; }
 	    void setAuto(bool a){ isAuto = a; }
@@ -51,15 +51,15 @@ namespace CameraManager {
         bool getCanAuto(){ return canAuto; }
         bool getOnOff(){ return onOff; }
 	    void setValue(float value){ val = value; }
-	    float getValue(){ return val; }
-	    float getDecimals(){ return decimals; }
+        double getValue() { return val; }
+        double getDecimals() { return decimals; }
 
 	    void setValueFromSlider(int value){ val = value/pow(10, (double) decimals); }
 	    int getValueToSlider(){ return val*pow(10, (double) decimals); }
 	    int getMinToSlider(){ return min*pow(10, (double) decimals); }
 	    int getMaxToSLider(){ return max*pow(10, (double) decimals); }
-        float getMin() { return min; }
-        float getMax() { return max; }
+        double getMin() { return min; }
+        double getMax() { return max; }
         QString getName() { return name; }
 	    QString formatValue(){ return QString::number(val, 'f', decimals); }
 
@@ -69,9 +69,9 @@ namespace CameraManager {
         bool canAuto;
         bool isAuto;
         bool onOff;
-        float min;
-        float max;
-        float val;
+        double min;
+        double max;
+        double val;
 	    int decimals;
 	};
 }
